backend/requirements.txt
alembic==1.14.1
bcrypt==4.3.0
blinker==1.8.2
charset-normalizer==3.4.2
click==8.1.8
defusedxml==0.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fpdf2==2.5.5
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
pillow==10.4.0
PyJWT==2.9.0
PyPDF2==1.26.0
python-dateutil==2.8.2
python-dotenv==1.0.1
reportlab==4.4.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
Faker==18.11.2
